refactor: divide cmd pkg and main func
