+++ cmd/go.mod
@@ -3,27 +3,18 @@ module github.com/siyul-park/uniflow/cmd
 go 1.21.0
 
 require (
-	github.com/go-faker/faker/v4 v4.2.0
-	github.com/jedib0t/go-pretty/v6 v6.4.9
-	github.com/oklog/ulid/v2 v2.1.0
-	github.com/pkg/errors v0.9.1
-	github.com/samber/lo v1.39.0
 	github.com/siyul-park/uniflow v0.0.0
+	github.com/siyul-park/uniflow/cmd/pkg v0.0.0
 	github.com/siyul-park/uniflow/plugin/pkg/controllx v0.0.0
 	github.com/siyul-park/uniflow/plugin/pkg/networkx v0.0.0
 	github.com/siyul-park/uniflow/plugin/pkg/systemx v0.0.0
-	github.com/spf13/cobra v1.8.0
 	github.com/spf13/viper v1.18.1
-	github.com/stretchr/testify v1.8.4
-	github.com/xiatechs/jsonata-go v1.8.4
 	go.mongodb.org/mongo-driver v1.13.1
-	gopkg.in/yaml.v3 v3.0.1
 )
 
 require (
 	github.com/acobaugh/osrelease v0.1.0 // indirect
 	github.com/benbjohnson/immutable v0.4.3 // indirect
-	github.com/davecgh/go-spew v1.1.2-0.20180830191138-d8f796af33cc // indirect
 	github.com/dlclark/regexp2 v1.7.0 // indirect
 	github.com/dop251/goja v0.0.0-20231027120936-b396bb4c349d // indirect
 	github.com/emirpasic/gods v1.18.1 // indirect
@@ -36,26 +27,31 @@ require (
 	github.com/hashicorp/hcl v1.0.0 // indirect
 	github.com/iancoleman/strcase v0.3.0 // indirect
 	github.com/inconshreveable/mousetrap v1.1.0 // indirect
+	github.com/jedib0t/go-pretty/v6 v6.4.9 // indirect
 	github.com/klauspost/compress v1.17.0 // indirect
 	github.com/magiconair/properties v1.8.7 // indirect
 	github.com/mattn/go-runewidth v0.0.13 // indirect
 	github.com/mitchellh/mapstructure v1.5.0 // indirect
 	github.com/montanaflynn/stats v0.7.1 // indirect
+	github.com/oklog/ulid/v2 v2.1.0 // indirect
 	github.com/pelletier/go-toml/v2 v2.1.0 // indirect
-	github.com/pmezard/go-difflib v1.0.1-0.20181226105442-5d4384ee4fb2 // indirect
+	github.com/pkg/errors v0.9.1 // indirect
 	github.com/rivo/uniseg v0.2.0 // indirect
 	github.com/sagikazarmark/locafero v0.4.0 // indirect
 	github.com/sagikazarmark/slog-shim v0.1.0 // indirect
+	github.com/samber/lo v1.39.0 // indirect
 	github.com/shopspring/decimal v1.3.1 // indirect
 	github.com/sourcegraph/conc v0.3.0 // indirect
 	github.com/spf13/afero v1.11.0 // indirect
 	github.com/spf13/cast v1.6.0 // indirect
+	github.com/spf13/cobra v1.8.0 // indirect
 	github.com/spf13/pflag v1.0.5 // indirect
 	github.com/subosito/gotenv v1.6.0 // indirect
 	github.com/tryvium-travels/memongo v0.10.0 // indirect
 	github.com/xdg-go/pbkdf2 v1.0.0 // indirect
 	github.com/xdg-go/scram v1.1.2 // indirect
 	github.com/xdg-go/stringprep v1.0.4 // indirect
+	github.com/xiatechs/jsonata-go v1.8.4 // indirect
 	github.com/youmark/pkcs8 v0.0.0-20201027041543-1326539a0a0a // indirect
 	go.uber.org/atomic v1.9.0 // indirect
 	go.uber.org/multierr v1.9.0 // indirect
@@ -65,9 +61,15 @@ require (
 	golang.org/x/sys v0.15.0 // indirect
 	golang.org/x/text v0.14.0 // indirect
 	gopkg.in/ini.v1 v1.67.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
 
 replace github.com/siyul-park/uniflow => ../
+
 replace github.com/siyul-park/uniflow/plugin/pkg/controllx => ../plugin/pkg/controllx
+
 replace github.com/siyul-park/uniflow/plugin/pkg/networkx => ../plugin/pkg/networkx
+
 replace github.com/siyul-park/uniflow/plugin/pkg/systemx => ../plugin/pkg/systemx
+
+replace github.com/siyul-park/uniflow/cmd/pkg => ./pkg
