+++ cmd/pkg/go.mod
@@ -1,0 +1,37 @@
+module github.com/siyul-park/uniflow/cmd/pkg
+
+go 1.21.0
+
+require (
+	github.com/go-faker/faker/v4 v4.2.0
+	github.com/jedib0t/go-pretty/v6 v6.4.9
+	github.com/oklog/ulid/v2 v2.1.0
+	github.com/pkg/errors v0.9.1
+	github.com/samber/lo v1.39.0
+	github.com/siyul-park/uniflow v0.0.0
+	github.com/spf13/cobra v1.8.0
+	github.com/stretchr/testify v1.8.4
+	github.com/xiatechs/jsonata-go v1.8.4
+	gopkg.in/yaml.v3 v3.0.1
+)
+
+require (
+	github.com/benbjohnson/immutable v0.4.3 // indirect
+	github.com/davecgh/go-spew v1.1.2-0.20180830191138-d8f796af33cc // indirect
+	github.com/emirpasic/gods v1.18.1 // indirect
+	github.com/goccy/go-json v0.10.2 // indirect
+	github.com/iancoleman/strcase v0.3.0 // indirect
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
+	github.com/kr/text v0.2.0 // indirect
+	github.com/mattn/go-runewidth v0.0.13 // indirect
+	github.com/pmezard/go-difflib v1.0.1-0.20181226105442-5d4384ee4fb2 // indirect
+	github.com/rivo/uniseg v0.2.0 // indirect
+	github.com/rogpeppe/go-internal v1.9.0 // indirect
+	github.com/shopspring/decimal v1.3.1 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
+	golang.org/x/exp v0.0.0-20231006140011-7918f672742d // indirect
+	golang.org/x/sys v0.15.0 // indirect
+	golang.org/x/text v0.14.0 // indirect
+)
+
+replace github.com/siyul-park/uniflow => ../../
