+++ cmd/pkg/resource/yamljson.go
