+++ plugin/pkg/controllx/coverage.coverprofile
@@ -1,0 +1,68 @@
+mode: atomic
+github.com/siyul-park/uniflow/plugin/pkg/controllx/builder.go:9.47,10.38 1 1
+github.com/siyul-park/uniflow/plugin/pkg/controllx/builder.go:10.38,12.105 2 1
+github.com/siyul-park/uniflow/plugin/pkg/controllx/builder.go:12.105,14.4 1 0
+github.com/siyul-park/uniflow/plugin/pkg/controllx/builder.go:16.3,17.102 2 1
+github.com/siyul-park/uniflow/plugin/pkg/controllx/builder.go:17.102,19.33 2 0
+github.com/siyul-park/uniflow/plugin/pkg/controllx/builder.go:19.33,20.49 1 0
+github.com/siyul-park/uniflow/plugin/pkg/controllx/builder.go:20.49,23.6 2 0
+github.com/siyul-park/uniflow/plugin/pkg/controllx/builder.go:25.4,25.17 1 0
+github.com/siyul-park/uniflow/plugin/pkg/controllx/builder.go:28.3,28.13 1 1
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:55.62,56.15 1 5
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:56.15,56.32 1 5
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:58.2,59.16 2 5
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:59.16,61.3 1 0
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:63.2,66.15 3 5
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:69.108,73.22 3 4
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:73.22,75.3 1 4
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:77.2,77.45 1 4
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:77.45,79.3 1 0
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:79.8,79.73 1 4
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:79.73,81.3 1 0
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:81.8,83.3 1 4
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:86.65,87.14 1 5
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:88.16,90.60 2 2
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:90.60,92.4 1 0
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:94.3,94.41 1 2
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:94.41,96.4 1 1
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:97.38,98.29 1 2
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:98.29,102.30 2 2
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:102.30,104.39 2 0
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:104.39,106.6 1 0
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:107.5,107.53 1 0
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:109.4,109.30 1 2
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:111.3,112.17 2 2
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:112.17,114.4 1 0
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:116.3,117.51 2 2
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:117.51,119.4 1 0
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:120.3,120.56 1 2
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:120.56,122.4 1 0
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:124.3,125.20 1 2
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:125.20,130.5 4 2
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:133.3,133.41 1 2
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:133.41,138.11 4 2
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:138.11,140.5 1 0
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:142.4,142.78 1 2
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:142.78,144.5 1 0
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:144.10,146.5 1 2
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:148.19,150.17 2 1
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:150.17,152.4 1 0
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:154.3,154.41 1 1
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:154.41,155.52 1 1
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:155.52,157.5 1 0
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:157.10,159.5 1 1
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:161.10,162.56 1 0
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:168.81,170.2 1 0
+github.com/siyul-park/uniflow/plugin/pkg/controllx/snippet.go:172.77,174.2 1 0
+github.com/siyul-park/uniflow/plugin/pkg/controllx/switch.go:46.34,51.2 3 2
+github.com/siyul-park/uniflow/plugin/pkg/controllx/switch.go:54.58,59.16 4 4
+github.com/siyul-park/uniflow/plugin/pkg/controllx/switch.go:59.16,61.3 1 0
+github.com/siyul-park/uniflow/plugin/pkg/controllx/switch.go:63.2,64.12 2 4
+github.com/siyul-park/uniflow/plugin/pkg/controllx/switch.go:68.36,74.2 4 2
+github.com/siyul-park/uniflow/plugin/pkg/controllx/switch.go:76.109,83.22 5 4
+github.com/siyul-park/uniflow/plugin/pkg/controllx/switch.go:83.22,85.3 1 4
+github.com/siyul-park/uniflow/plugin/pkg/controllx/switch.go:87.2,87.36 1 4
+github.com/siyul-park/uniflow/plugin/pkg/controllx/switch.go:87.36,88.93 1 10
+github.com/siyul-park/uniflow/plugin/pkg/controllx/switch.go:88.93,89.59 1 4
+github.com/siyul-park/uniflow/plugin/pkg/controllx/switch.go:89.59,94.5 3 4
+github.com/siyul-park/uniflow/plugin/pkg/controllx/switch.go:98.2,98.62 1 0
